fix - kya widgets
